Crate solana_streamer

Modules

evicting_sender

nonblocking

packet

The packet module defines data structures and methods to pull data from the network.

quic

recvmmsg

The recvmmsg module provides recvmmsg() API implementation

sendmmsg

The sendmmsg module provides sendmmsg() API implementation

socket

streamer

The streamer module defines a set of services for efficiently pulling data from UDP sockets.