solana_streamer::nonblocking

Module quic

Structs

ClientConnectionTracker

struct ease tracking connections of all stages, so that we do not have to litter the code with open connection tracking. This is added into the connection table as part of the ConnectionEntry. The reference is auto reduced when it is dropped.

SpawnNonBlockingServerResult

Enums

ConnectionPeerType

Constants

ALPN_TPU_PROTOCOL_ID

DEFAULT_WAIT_FOR_CHUNK_TIMEOUT

Functions

get_connection_stake

get_remote_pubkey

spawn_serverDeprecated

spawn_server_multiDeprecated

spawn_server_with_cancel

Spawn a streamer instance in the current tokio runtime.