solana_streamer

Module quic

Structs

EndpointKeyUpdater

QuicServerParamsDeprecated

QuicStreamerConfig

SimpleQosQuicStreamerConfig

SpawnServerResult

StreamerStats

SwQosQuicStreamerConfig

Enums

QuicServerError

Constants

DEFAULT_MAX_CONNECTIONS_PER_IPADDR_PER_MINUTE

The new connections per minute from a particular IP address. Heuristically set to the default maximum concurrent connections per IP address. Might be adjusted later.

DEFAULT_MAX_QUIC_CONNECTIONS_PER_PEER

DEFAULT_MAX_STAKED_CONNECTIONS

DEFAULT_MAX_STREAMS_PER_MS

Limit to 500K PPS

DEFAULT_MAX_UNSTAKED_CONNECTIONS

DEFAULT_QUIC_ENDPOINTS

Functions

default_num_tpu_transaction_forward_receive_threads

default_num_tpu_transaction_receive_threads

default_num_tpu_vote_transaction_receive_threads

spawn_serverDeprecated

spawn_server_multiDeprecated

spawn_server_with_cancel

Spawns a tokio runtime and a streamer instance inside it.

spawn_simple_qos_server_with_cancel

Spawns a tokio runtime and a streamer instance inside it.